added exception handling and service layer

+++ src/main/java/com/practice/videogametracekrservice/VideoGameService.java
@@ -22,6 +22,9 @@ public class VideoGameService {
     }
 
     public void deleteVideoGame(Long id) {
+        if(!vgRepos.existsById(id)) {
+            throw new RuntimeException("Game with id " + id + " does not exist");
+        }
         vgRepos.deleteById(id);
     }
